alterando service e controller para funfar com conexao com mysql

+++ SWAPI_ReactJS_Spring/api/swapi/src/main/java/br/com/cocayan/swapi/controllers/PeopleController.java
@@ -1,5 +1,8 @@
 package br.com.cocayan.swapi.controllers;
 
+import java.net.URI;
+import java.util.Optional;
+
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +10,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,6 +19,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import br.com.cocayan.swapi.entities.People;
 import br.com.cocayan.swapi.entities.dtos.CreatePeopleDto;
@@ -38,61 +43,57 @@ public class PeopleController {
             size = 10
         ) Pageable pageable
     ) {
-        Page<People> peoples = peopleService.getAllPeople(pageable);
+        Page<People> peoples = peopleService.getAllPeoples(pageable);
         return PeopleDto.pagePeopleToPagePeopleDto(peoples);
     }
 
     @GetMapping("/{peopleId}")
-    public String getPeopleById(@PathVariable Long peopleId) {
-        try {
-            PeopleDto peopleDto = PeopleDto.convertPeopleToPeopleDto(peopleService.getPeopleById(peopleId));
-            
-            if (peopleDto != null) {
-                return peopleDto.toString();
-            } else {
-                return "People com o id: " + peopleId + ", não foi encontrado...";
-            }
-        } catch (Exception e) {
-            System.err.println(e);
-            return "People com o id: " + peopleId + ", não foi encontrado...";
+    public ResponseEntity<PeopleDto> getPeopleById(@PathVariable Long peopleId) {
+        Optional<People> optional = peopleService.getPeopleById(peopleId);
+
+        if (optional.isPresent()) {
+            return ResponseEntity.ok(new PeopleDto(optional.get()));
+        } else {
+            return ResponseEntity.notFound().build();
         }
     }
 
     @PostMapping
-    public PeopleDto createPeople(@RequestBody @Valid CreatePeopleDto createPeopleDto) {
-        People people = new People(
-            createPeopleDto.getName(), 
-            createPeopleDto.getHeight(), 
-            createPeopleDto.getMass()
-        );
+    public ResponseEntity<PeopleDto> createPeople(
+        @RequestBody @Valid CreatePeopleDto createPeopleDto, 
+        UriComponentsBuilder uriBuilder
+    ) {
+        People people = createPeopleDto.createPeopleDtoToPeople(createPeopleDto);
+        peopleService.createPeople(people);
 
-        return PeopleDto.convertPeopleToPeopleDto(peopleService.createPeople(people));
+        URI uri = uriBuilder.path("people/{peopleId}").buildAndExpand(people.getPeopleId()).toUri();
+        return ResponseEntity.created(uri).body(new PeopleDto(people));
     }
 
     @PutMapping
-    public String updatePeople(@RequestBody @Valid UpdatePeopleDto updatePeopleDto) {
-        People people = new People(
-            updatePeopleDto.getPeopleId(), 
-            updatePeopleDto.getName(), 
-            updatePeopleDto.getHeight(), 
-            updatePeopleDto.getMass()
-        );
+    public ResponseEntity<PeopleDto> updatePeople(@RequestBody @Valid UpdatePeopleDto updatePeopleDto) {
+        Optional<People> optional = peopleService.getPeopleById(updatePeopleDto.getPeopleId());
 
-        PeopleDto peopleDto = PeopleDto.convertPeopleToPeopleDto(peopleService.updatePeople(people));
-
-        if (peopleDto != null) {
-            return peopleDto.toString();
+        if (optional.isPresent()) {
+            People people = updatePeopleDto.updatePeopleDtoToPeople(updatePeopleDto);
+            return ResponseEntity.ok(new PeopleDto(peopleService.updatePeople(people)));
         } else {
-            return "People com o id: " + updatePeopleDto.getPeopleId() + ", não foi encontrado...";
-        }    
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping
-    public String deletePeople(@RequestBody PeopleDto peopleDto) {
-        if (peopleService.deletePeople(peopleDto.getPeopleId())) {
-            return "People com o id: " + peopleDto.getPeopleId() + ", deletado com sucesso!";
+    public ResponseEntity<?> deletePeople(@RequestBody PeopleDto peopleDto) {
+        Optional<People> optional = peopleService.getPeopleById(peopleDto.getPeopleId());
+
+        if (optional.isPresent()) {
+            if (peopleService.deletePeople(peopleDto.getPeopleId())) {
+                return ResponseEntity.ok().build();
+            } else {
+                return ResponseEntity.noContent().build();
+            }
         } else {
-            return "People com o id: " + peopleDto.getPeopleId() + ", não foi encontrado...";
+            return ResponseEntity.notFound().build();
         }
     }
